Adjust token bucket setup.
Originally this was configured to continually fill a bucket at a rate of
x per second. As such, a violator would have to spike a large number of
requests to get immediately caught. Those hovering just over the
ratelimit wouldn't exhaust the bucket for a long time due to tokens
continually refilling.

This change initializes the bucket and only refills it once every
defined interval. For example, if you define a limit of 180 requests
over 3 minutes, 180 tokens would be inserted at each 3 minute interval.
This ensures that those just barely violating the ratelimit are limited
in a timely manner.

This change is not perfect - it still allows abusers to spike 2x the
limit if the coordinate around the refill. (1x before the token refill,
and 1x immediately after).